deploy/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
temp
pdfminer
pdfminer.six
flask
fitz